move suggestions to its own package
