+++ Sources/Suggestions/APIResult.swift
