+++ Sources/Suggestions/SuggestionResult.swift
