+++ Tests/SuggestionsTests/APIResultTests.swift
@@ -17,7 +17,7 @@
 //
 
 import XCTest
-@testable import BrowserServicesKit
+@testable import Suggestions
 
 final class APIResultTests: XCTestCase {
